+++ cli/src/package_managers/cargo.rs
@@ -4,10 +4,10 @@ use crate::parsers::toml;
 #[derive(Debug)]
 pub enum CargoTomlError {
     InvalidPackageFieldDataType { workspace: bool },
-    InvalidVersionFieldDataType { workspace: bool },
+    InvalidPackageVersionFieldDataType { workspace: bool },
     InvalidWorkspaceFieldDataType,
     MissingPackageField { workspace: bool },
-    MissingVersionField { workspace: bool },
+    MissingPackageVersionField { workspace: bool },
     ParseToml(Box<toml_edit::TomlError>),
 }
 
@@ -18,7 +18,7 @@ impl core::fmt::Display for CargoTomlError {
     fn fmt(&self, f: &mut core::fmt::Formatter<'_>) -> core::fmt::Result {
         match self {
             Self::ParseToml(error) => error.fmt(f),
-            Self::InvalidVersionFieldDataType { workspace } => {
+            Self::InvalidPackageVersionFieldDataType { workspace } => {
                 let field = if *workspace {
                     "\"workspace.package.version\""
                 } else {
@@ -36,7 +36,7 @@ impl core::fmt::Display for CargoTomlError {
 
                 write!(f, "{field} field is not a table")
             }
-            Self::MissingVersionField { workspace } => {
+            Self::MissingPackageVersionField { workspace } => {
                 let field = if *workspace {
                     "\"workspace.package.version\""
                 } else {
@@ -67,11 +67,11 @@ fn set_package_version(
 ) -> Result<bool, CargoTomlError> {
     let version_key = package_table
         .get("version")
-        .ok_or(CargoTomlError::MissingVersionField { workspace })?;
+        .ok_or(CargoTomlError::MissingPackageVersionField { workspace })?;
 
     let version_key_str = version_key
         .as_str()
-        .ok_or(CargoTomlError::InvalidVersionFieldDataType { workspace })?;
+        .ok_or(CargoTomlError::InvalidPackageVersionFieldDataType { workspace })?;
 
     let modified = version_key_str != version;
 
@@ -144,6 +144,7 @@ pub fn update_lock_files(dir: &std::path::Path) -> std::io::Result<bool> {
 
 #[cfg(test)]
 mod test_set_cargo_toml_version {
+    use super::{CargoTomlError, set_cargo_toml_version};
     use crate::package_managers::error::PackageManagerError;
 
     #[test]
@@ -176,7 +177,7 @@ toml_edit = "0.22.26"
         assert!(expected_output.contains(&new_version_line));
 
         let (modified, output) =
-            super::set_cargo_toml_version(input.to_string(), version).expect("it not to raise");
+            set_cargo_toml_version(input.to_string(), version).expect("it not to raise");
 
         assert!(modified);
 
@@ -185,7 +186,7 @@ toml_edit = "0.22.26"
         // Validate we do not modify file if version is the same
         {
             let (modified, output) =
-                super::set_cargo_toml_version(output, version).expect("it not to raise");
+                set_cargo_toml_version(output, version).expect("it not to raise");
 
             assert!(!modified);
 
@@ -230,7 +231,7 @@ toml_edit = "0.22.26"
         assert!(expected_output.contains(&new_version_line));
 
         let (modified, output) =
-            super::set_cargo_toml_version(input.to_string(), version).expect("it not to raise");
+            set_cargo_toml_version(input.to_string(), version).expect("it not to raise");
 
         assert!(modified);
 
@@ -239,7 +240,7 @@ toml_edit = "0.22.26"
         // Validate we do not modify file if version is the same
         {
             let (modified, output) =
-                super::set_cargo_toml_version(output, version).expect("it not to raise");
+                set_cargo_toml_version(output, version).expect("it not to raise");
 
             assert!(!modified);
 
@@ -251,101 +252,172 @@ toml_edit = "0.22.26"
     fn it_should_require_package_field() {
         let input = "";
 
-        let result = super::set_cargo_toml_version(input.to_string(), "1.23.4")
+        let result = set_cargo_toml_version(input.to_string(), "1.23.4")
             .expect_err("it should return an error");
 
         assert!(matches!(
             result,
-            super::CargoTomlError::MissingPackageField { workspace: false }
+            CargoTomlError::MissingPackageField { workspace: false }
         ));
 
-        assert!(result.to_string().contains("\"package\""));
-
-        PackageManagerError::from(result).test_up_casting();
+        assert!(
+            crate::error::Error::from(PackageManagerError::from(result))
+                .to_string()
+                .contains("\"package\"")
+        );
     }
 
     #[test]
     fn it_should_require_package_version_field() {
         let input = "[package]";
 
-        let result = super::set_cargo_toml_version(input.to_string(), "1.23.4")
+        let result = set_cargo_toml_version(input.to_string(), "1.23.4")
             .expect_err("it should return an error");
 
         assert!(matches!(
             result,
-            super::CargoTomlError::MissingVersionField { workspace: false }
+            CargoTomlError::MissingPackageVersionField { workspace: false }
         ));
 
-        assert!(result.to_string().contains("\"package.version\""));
-
-        PackageManagerError::from(result).test_up_casting();
+        assert!(
+            crate::error::Error::from(PackageManagerError::from(result))
+                .to_string()
+                .contains("\"package.version\"")
+        );
     }
 
     #[test]
     fn workspace_should_require_package_field() {
         let input = "[workspace]";
 
-        let result = super::set_cargo_toml_version(input.to_string(), "1.23.4")
+        let result = set_cargo_toml_version(input.to_string(), "1.23.4")
             .expect_err("it should return an error");
 
         assert!(matches!(
             result,
-            super::CargoTomlError::MissingPackageField { workspace: true }
+            CargoTomlError::MissingPackageField { workspace: true }
         ));
 
-        assert!(result.to_string().contains("\"workspace.package\""));
-
-        PackageManagerError::from(result).test_up_casting();
+        assert!(
+            crate::error::Error::from(PackageManagerError::from(result))
+                .to_string()
+                .contains("\"workspace.package\"")
+        );
     }
 
     #[test]
     fn workspace_should_require_package_version_field() {
         let input = "[workspace.package]";
 
-        let result = super::set_cargo_toml_version(input.to_string(), "1.23.4")
+        let result = set_cargo_toml_version(input.to_string(), "1.23.4")
             .expect_err("it should return an error");
 
         assert!(matches!(
             result,
-            super::CargoTomlError::MissingVersionField { workspace: true }
+            CargoTomlError::MissingPackageVersionField { workspace: true }
         ));
 
-        assert!(result.to_string().contains("\"workspace.package.version\""));
-
-        PackageManagerError::from(result).test_up_casting();
+        assert!(
+            crate::error::Error::from(PackageManagerError::from(result))
+                .to_string()
+                .contains("\"workspace.package.version\"")
+        );
     }
 
     #[test]
     fn package_should_be_map() {
         let input = "package = \"123\"\n";
 
-        let result = super::set_cargo_toml_version(input.to_string(), "1.23.4")
+        let result = set_cargo_toml_version(input.to_string(), "1.23.4")
             .expect_err("it should return an error");
 
         assert!(matches!(
             result,
-            super::CargoTomlError::InvalidPackageFieldDataType { workspace: false }
+            CargoTomlError::InvalidPackageFieldDataType { workspace: false }
         ));
 
-        assert!(result.to_string().contains("\"package\""));
+        assert!(
+            crate::error::Error::from(PackageManagerError::from(result))
+                .to_string()
+                .contains("\"package\"")
+        );
+    }
 
-        PackageManagerError::from(result).test_up_casting();
+    #[test]
+    fn package_version_should_be_string() {
+        let input = "[package.version]\nkey = \"123\"\n";
+
+        let result = set_cargo_toml_version(input.to_string(), "1.23.4")
+            .expect_err("it should return an error");
+
+        assert!(matches!(
+            result,
+            CargoTomlError::InvalidPackageVersionFieldDataType { workspace: false }
+        ));
+
+        assert!(
+            crate::error::Error::from(PackageManagerError::from(result))
+                .to_string()
+                .contains("\"package.version\"")
+        );
+    }
+
+    #[test]
+    fn workspace_should_be_map() {
+        let input = "workspace = \"123\"\n";
+
+        let result = set_cargo_toml_version(input.to_string(), "1.23.4")
+            .expect_err("it should return an error");
+
+        assert!(matches!(
+            result,
+            CargoTomlError::InvalidWorkspaceFieldDataType
+        ));
+
+        assert!(
+            crate::error::Error::from(PackageManagerError::from(result))
+                .to_string()
+                .contains("\"workspace\"")
+        );
     }
 
     #[test]
     fn workspace_package_should_be_map() {
         let input = "[workspace]\npackage = \"123\"\n";
 
-        let result = super::set_cargo_toml_version(input.to_string(), "1.23.4")
+        let result = set_cargo_toml_version(input.to_string(), "1.23.4")
             .expect_err("it should return an error");
 
         assert!(matches!(
             result,
-            super::CargoTomlError::InvalidPackageFieldDataType { workspace: true }
+            CargoTomlError::InvalidPackageFieldDataType { workspace: true }
         ));
 
-        assert!(result.to_string().contains("\"workspace.package\""));
+        assert!(
+            crate::error::Error::from(PackageManagerError::from(result))
+                .to_string()
+                .contains("\"workspace.package\"")
+        );
+    }
 
-        PackageManagerError::from(result).test_up_casting();
+    #[test]
+    fn workspace_package_version_should_be_string() {
+        let input = "[workspace.package.version]\nkey = \"123\"\n";
+
+        let result = set_cargo_toml_version(input.to_string(), "1.23.4")
+            .expect_err("it should return an error");
+
+        dbg!(&result);
+
+        assert!(matches!(
+            result,
+            CargoTomlError::InvalidPackageVersionFieldDataType { workspace: true }
+        ));
+
+        assert!(
+            crate::error::Error::from(PackageManagerError::from(result))
+                .to_string()
+                .contains("\"workspace.package.version\"")
+        );
     }
 }