+++ cli/src/package_managers/npm.rs
@@ -112,12 +112,11 @@ pub fn update_lock_files(dir: &std::path::Path) -> std::io::Result<bool> {
 
 #[cfg(test)]
 mod test_set_package_json_version {
+    use super::{PackageJsonError, set_package_json_version};
     use crate::package_managers::error::PackageManagerError;
 
-    use super::{PackageJsonError, set_package_json_version};
-
     #[test]
-    fn it_should_modify_version() -> Result<(), super::PackageJsonError> {
+    fn it_should_modify_version() {
         let version = "1.2.3";
 
         let input = "{
@@ -137,7 +136,8 @@ mod test_set_package_json_version {
 
         assert!(expected_output.contains(&new_version_line));
 
-        let (modified, output) = set_package_json_version(input.to_string(), version)?;
+        let (modified, output) =
+            set_package_json_version(input.to_string(), version).expect("it not to raise");
 
         assert!(modified);
 
@@ -145,14 +145,13 @@ mod test_set_package_json_version {
 
         // Validate we do not modify file if version is the same
         {
-            let (modified, output) = set_package_json_version(output, version)?;
+            let (modified, output) =
+                set_package_json_version(output, version).expect("it not to raise");
 
             assert!(!modified);
 
             assert_eq!(output, expected_output);
         }
-
-        Ok(())
     }
 
     #[test]
@@ -164,9 +163,11 @@ mod test_set_package_json_version {
 
         assert!(matches!(result, PackageJsonError::MissingVersionField));
 
-        assert!(result.to_string().contains("\"version\""));
-
-        PackageManagerError::from(result).test_up_casting();
+        assert!(
+            crate::error::Error::from(PackageManagerError::from(result))
+                .to_string()
+                .contains("\"version\"")
+        );
     }
 
     #[test]
@@ -181,8 +182,10 @@ mod test_set_package_json_version {
             PackageJsonError::InvalidVersionFieldDataType
         ));
 
-        assert!(result.to_string().contains("\"version\""));
-
-        PackageManagerError::from(result).test_up_casting();
+        assert!(
+            crate::error::Error::from(PackageManagerError::from(result))
+                .to_string()
+                .contains("\"version\"")
+        );
     }
 }